Update WebScreenGrabV2.py
Column Name Change:
The header in the Excel output has been updated from “IP/Host” to “Ip Address” and the corresponding key in row_data is now "ip_address".

Removal of XML and CSV:
All functions and references to XML and CSV output have been removed. The script now solely processes data and writes the results to an Excel file.